Support for firmware managed devices
This change adds a flag to the device structure such that the devices
can present themselves as firmware managed devices. Firmware managed
devices are conceptually guest visible devices that require some
coordination with the OpenHCL firmware to be usable. For example, MANA
devices that require netvsp to be operational are considered firmware
managed devices.