netvsp: always wait for coordinator to process messages from the primary worker (#2370)
There have been several issues with starting the primary worker before
the last message was processed (specifically the message to switch the
data path). These have been resolved as discovered, but there is still
at least one race, and generally this should be made harder to get
wrong. Change the protocol to put the primary worker in a "waiting"
state, and only after the message has been processed is it returned to
'ready'. The state is used to determine if it should be restarted, so it
should never actually start running in the 'waiting' state, but just in
case that will pend instead of running the main_loop.

I was not able to hit the latest race condition via unit test, but the
test code has been updated with fixes and behavior updates that more
closely match real devices.

Co-authored-by: Brian Perkins <brian.perkins@microsoft.com>