vmbus_server: remove incomplete GPADLs when a channel is released (#2410)
If a channel was released (by the guest, or when the connection is reset
due to a guest crash or VM reset) while it had an in-progress GPADL, we
would not remove the GPADL ID from the `incomplete_gpadls` map. This
would cause a panic if the guest tried to reuse the same GPADL ID later.

This change makes sure that incomplete GPADLs are properly removed from
that list.

It also updates the logic for detecting duplicate GPADL IDs, so in case
a guest actually does use a duplicate, it doesn't cause a panic. The
existing logic allowed duplicate gpadl IDs for different channels to
handle reserved channels, but would panic on duplicate in-progress IDs.

A test is added for this behavior.