nvme_manager: don't reset devices after first restore (#2416)
The fresh create path prefers to use the `NoReset` method when creating
devices, this prevents
the device from receiving a reset when the VFIO handle is dropped. The
restore path does not do this.

This implies that the device will see an FLR on the second save/restore.

Fix that, and also add some verbose logging that helped track this down.