enable default_boot_always_attempt for non-TL VMs (#2436)
Always enable default boot always attempt for non-Trusted Launch VMs.
This is roughly equivalent to not having secure boot or TPM enabled.
This is necessary because the VMGS is not swapped with the OS disk for
these VMs in Azure (and in any case on-prem), causing the VM to fail to
boot after an OS swap.