build_info: extend BuildInfo with extra fields when built locally (#2454)
This change adds additional fields to the BuildInfo struct when OpenHCL
is built locally. Here's an example output of `x -v build_info` from the
OpenVMM interactive console:
```
{
    author_email: "campjustus@microsoft.com",
    author_name: "Justus Camp",
    build_machine: "DESKTOP-J6M7TP0",
    build_timestamp: "Wed, 19 Nov 2025 12:49:08 -0800",
    built_with_uncommitted_changes: "true",
    crate_name: "build_info",
    internal_scm_branch: "",
    internal_scm_revision: "",
    openhcl_version: "",
    scm_branch: "HEAD",
    scm_revision: "f7e6eddf938f324e394ad16b6ce7d13714e9caae",
}
```

This does add a new dep, the "hostname" crate which handles grabbing the
hostname cross-platform. I can implement this locally instead if we
don't want to introduce a new dep.