GET: Make the client MeshPayload (#2464)
As part of my work to support running ChipsetDevices in another process,
I ran into needing certain resources to be MeshPayload. The GET client
is the first of these. The framebuffer is another one I initially hadn't
planned on doing, but turns out it's free once the GET is done, so add
it too.