flowey: remove echo for bootstrap flowey (#2474)
This echo isn't adding anything valuable and is actually causing some
issues with our internal ADO pipelines. For some additional context we
have some Windows CI images that do not have the `bash.exe` in the path,
which makes this task fail. I'm working on getting those images fixed,
but in the meantime, this will allow the bootstrapping to succeed on
these images.

Co-authored-by: OpenVMM Team <openvmm@microsoft.com>