petri: wait for shutdown ic before pipette (#2527)
As a workaround for #2470 (where the guest crashes when the pipette
connection timeout expires due to a vmbus bug), wait for the shutdown IC
to come online first so that we probably won't time out when connecting
to the agent. Undoes https://github.com/microsoft/openvmm/pull/2485 in
hopes that this is sufficient.