logview_new: update packages (#2538)
Updating react and react-dom packages to the latest ones (Along with the
typescript definitions).
Including a manual build for this one because that had been built with
older versions. However, ci is already building the website now. Once
node version is upgraded in CI, the manually built version will be
removed completely