nvme_driver: Avoid using CancelContext on hot path (#2426)
As per https://microsoft.visualstudio.com/OS/_workitems/edit/60178470,
`CancelContext` as used here is polled every time the async task is
woken up. This is causing high CPU usage on a hot CPU in local
performance testing. This PR changes it to `TaskControl`.

The performance improvement, if any, is still being validated, but we
want this change in either case.