logview_new: safely set inspect links in the log viewer (#2522)
Previously we would treat the entire message as HTML so that it can
display links to the inspect viewer. This PR fixes it by separating the
links from the actual message. As a side effect it also fixes an issue
where highlighted text would get refreshed/unhighlighted when the page
was scrolled.