flowey: Remove support for building tests with panic-abort-tests (#2558)
This removes yet another use of RUSTC_BOOTSTRAP from our CI, which is
something we should be striving towards. We don't even use this feature
anymore in this repo. There was one use in the internal repo, however a
test run I did there seems to indicate that this is no longer needed
since we switched to nextest. So we can safely get rid of it.