proxyintegration: Re-implement user-mode NUMA node map (#2560)
We decided adding more kernel mode code wasn't the way to go, so this
change delegates fetching the NUMA configuration to user-mode VMBus.
Instead of calling a proxy IOCTL on the task start, we now accept a NUMA
node map from user-mode VMBus in the proxy builder, which is passed down
into the `ProxyTask`.