hcl: Rewrite get/set vp register paths (#2592)
This is a major rewrite of all of the get/set VP register paths
previously contained in ioctl.rs. The goal is to make the code easier to
understand and follow, reduce duplicated checks and paths, reduce
guest_arch splits, and move it into a dedicated file to clean up
ioctl.rs, with minimal changes to consumers. Hopefully I've accomplished
all that.

As future work we may want to consider removing the ability to access
shared registers completely from this code, and instead require backings
to access their shared_context/vmsa/vmcs directly as appropriate.

Also as a drive-by move arm's translate_gva_to_gpa into the arm-specific
file to match x64's, instead of having it just live in ioctl.rs.