netvsp: handle eqe 135 and reconfigure vf (#2576)
For netvsp to recover from SoC crash and NMC servicing, it must detect
EQE 135 from MANA and reconfigure the Virtual Function.

* `GDMA_EQE_HWC_RECONFIG_VF` added to events handled by the GDMA driver
* `vf_reconfiguration_pending` bool added the GDMA driver. Set when the
EQE is received, read by the MANA driver when processing all EQ events.
* `vf_reconfig_sender` added to MANA driver to signal Netvsp VF Manager
when it sees 'pending' is true
* `vf_reconfig_receiver` added to Netvsp HclNetworkVFManagerWorker to
send `VFReconfig` message when signaled
* `VfReconfig` message added to Netvsp VF Manager, which removes the old
VF and then creates a new VF

Smaller changes:
* `GDMA_GENERATE_TEST_EQE` and `GDMA_GENERATE_RECONFIG_VF_EVENT` added
in order to create a unit test, `test_gdma_reconfig_vf()`
* The logic of `NextWorkItem::ManaDeviceArrived` refactored into
`startup_vtl2_device()` so the logic can be shared with `VfReconfig`

Testing:
* Unit tests pass
* Tested on a lab machine with SoC MANA privates which allowed EQE 135
to be generated by command. Netvsp is able to see the EQE and VfReconfig
is called. Before and after sending the EQE, ping and ntttcp traffic
succeed.